[신아름] 적어둔 메모에 추천/ 비추천/ 삭제 구현 [CAL-2]

+++ AreumSin/backend/demo/src/main/java/com/example/demo/repository/JoinMemberRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import javax.transaction.Transactional;
 import java.util.Optional;
 
-public interface JoinMemberRepository extends JpaRepository<Member, Long> {
+public interface JoinMemberRepository extends JpaRepository<Member, Integer> {
     Boolean existsById(String id);
 
     @Transactional
